Hay algo mal pero funciona

+++ Fusion_System/Fusion_System.u/local_interfaces/Match_Objects.h
@@ -4,7 +4,7 @@ typedef struct {
 	int number_matched[AUTO_MAX_NUM_OBJECTS];//Vector que indica numero de objetos asociados a cada uno
 	int Matrix_matched[AUTO_MAX_NUM_OBJECTS][AUTO_MAX_NUM_OBJECTS];//Matriz con todos los objetos
 	//TODO::en pruebas
-	int overlap[AUTO_MAX_NUM_OBJECTS][AUTO_MAX_NUM_OBJECTS][2];//% de coincidencia pos 0 % de coincidencia real pos 1
+	float overlap[AUTO_MAX_NUM_OBJECTS][AUTO_MAX_NUM_OBJECTS][2];//% de coincidencia pos 0 % de coincidencia real pos 1
 	//Primer parametro tamaño numer_objects 
 	//Segundo parametro lista de identificadores asociados a cada objeto
 }MATCH_OBJECTS;
